# docs/swagger.yaml
definitions:
  v1.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  v1.GamePopularity:
    properties:
      name:
        type: string
      popularity:
        type: integer
    type: object
  v1.ModlistSummaryResponse:
    properties:
      Name:
        type: string
      link:
        type: string
    type: object
  v1.RepositoryResponse:
    properties:
      link:
        type: string
      name:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for parsing Wabbajack modlists
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Wabbajack Modlist Parser API
  version: "1.0"
paths:
  /api/v1/games:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available games
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all games
      tags:
      - games
  /api/v1/games/top-popular:
    get:
      consumes:
      - application/json
      description: Retrieves a list of games ranked by popularity based on the number
        of modpacks/modlists available for each game. The popularity score is determined
        by counting how many modpacks reference each game across all repositories.
      parameters:
      - description: 'Number of top games to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Sort order: ''asc'' for ascending (least popular first), ''desc''
          for descending (most popular first) (default: ''desc'')'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of games with their popularity scores
          schema:
            items:
              $ref: '#/definitions/v1.GamePopularity'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "502":
          description: Service temporarily unavailable
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get top popular games
      tags:
      - games
  /api/v1/health:
    get:
      consumes:
      - application/json
      description: Check if the API is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
  /api/v1/modlists:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available modlists with their details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.ModlistSummaryResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all modlists
      tags:
      - modlists
  /api/v1/repositories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available repositories with their details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.RepositoryResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all repositories
      tags:
      - repositories
  /api/v1/status:
    get:
      consumes:
      - application/json
      description: Get API version and status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: API status
      tags:
      - api
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"